Put the default implementation of the ResultSet interface in a separated abstract class. Move the tests into the good files.
